backend/requirements.txt
certifi==2024.6.2
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
httpx==0.27.0
psycopg2==2.9.9
pydantic==2.8.0
pydantic-settings==2.3.4
pydantic_core==2.20.0
PyJWT==2.8.0
pytest==8.3.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
SQLAlchemy==2.0.31
uvicorn==0.30.1
alembic==1.13.1
